Reorg a couple tests to keep 1:1 relationship with defaults being tested.
